Move robotile scripts to Base, use symlinks
