+++ Projects/Base/Data/RoboTile/Bricks.xml
@@ -1,0 +1,5 @@
+<RoboTileSet>
+<Script Name="Fill" Priority="0">
+require("Bricks")
+</Script>
+</RoboTileSet>
