+++ Projects/Base/Data/RoboTile/Fungal.xml
@@ -1,0 +1,5 @@
+<RoboTileSet>
+<Script Name="Fill" Priority="0">
+require("Fungal")
+</Script>
+</RoboTileSet>
