+++ Projects/Base/Data/RoboTile/GreenBrinstar.xml
@@ -1,0 +1,5 @@
+<RoboTileSet>
+<Script Name="Fill" Priority="0">
+require("GreenBrinstar")
+</Script>
+</RoboTileSet>
