+++ Projects/Base/Data/RoboTile/OuterCrateria.xml
@@ -1,0 +1,5 @@
+<RoboTileSet>
+<Script Name="Fill" Priority="0">
+require("OuterCrateria")
+</Script>
+</RoboTileSet>
