+++ Projects/Base/Data/RoboTile/RedBrinstar.xml
@@ -1,0 +1,5 @@
+<RoboTileSet>
+<Script Name="Fill" Priority="0">
+require("RedBrinstar")
+</Script>
+</RoboTileSet>
