+++ Projects/Base/Data/RoboTile/WestMaridiaThick.xml
@@ -1,0 +1,5 @@
+<RoboTileSet>
+<Script Name="Fill" Priority="0">
+require("WestMaridiaThick")
+</Script>
+</RoboTileSet>
