+++ Projects/Base/Data/RoboTile/WestMaridiaThin.xml
@@ -1,0 +1,5 @@
+<RoboTileSet>
+<Script Name="Fill" Priority="0">
+require("WestMaridiaThin")
+</Script>
+</RoboTileSet>
