+++ Projects/Base/Data/RoboTile/WreckedShip.xml
@@ -1,0 +1,5 @@
+<RoboTileSet>
+<Script Name="Fill" Priority="0">
+require("WreckedShip")
+</Script>
+</RoboTileSet>
