+++ Projects/InnerCrateria/Data/RoboTile/Fungal.xml
@@ -1,5 +1,1 @@
-<RoboTileSet>
-<Script Name="Fill" Priority="0">
-require("Fungal")
-</Script>
-</RoboTileSet>
+../../../Base/Data/RoboTile/Fungal.xml
