+++ Projects/MechaTourian/Data/RoboTile/MechaTourianBlockFill.xml
@@ -1,5 +1,1 @@
-<RoboTileSet>
-<Script Name="Fill" Priority="0">
-require("MechaTourianBlockFill")
-</Script>
-</RoboTileSet>
+../../../Base/Data/RoboTile/MechaTourianBlockFill.xml
