--- Projects/OuterCrateria/Data/RoboTile/FixBGTiles.lua
@@ -1,9 +1,0 @@
-if t:gfx_tile(0, 0) == 0x13D then
-    t:set_gfx(0x111, false, false)
-end
-if t:gfx_tile(0, 0) == 0x13E then
-    t:set_gfx(0x112, false, false)
-end
-if t:gfx_tile(0, 0) == 0x13F then
-    t:set_gfx(0x113, false, false)
-end
