+++ Projects/OuterCrateria/Data/RoboTile/OuterCrateria.xml
@@ -1,5 +1,1 @@
-<RoboTileSet>
-<Script Name="Fill" Priority="0">
-require("OuterCrateria")
-</Script>
-</RoboTileSet>
+../../../Base/Data/RoboTile/OuterCrateria.xml
