+++ Projects/RedBrinstar/Data/RoboTile/RedBrinstar.lua
@@ -1,257 +1,1 @@
-require("Common")
-
-tile_air = 0x0FF
-tile_unknown = 0x0DF
-tile_unknown_solid = 0x162
-tile_interior = 0x165
-tile_bottom_edge = 0x104
-
-tile_top_edge = 0x124
-tile_left_edge = 0x103
-tile_bottom_left_outside_corner = 0x101
-tile_top_left_outside_corner = 0x121
-
-tile_bottom_right_inside_corner = 0x26C
-tile_top_right_inside_corner = 0x247
-
-tile_half_bottom_edge = 0x234
-tile_half_top_edge = 0x274
-tile_half_left_edge = 0x255
-
-tile_bottom_right_45_slope = 0x26B
-tile_under_bottom_right_45_slope = 0x28B
-
-tile_top_right_45_slope = 0x221
-tile_above_top_right_45_slope = 0x222
-
-tile_top_right_45_small_slope = 0x261   
-tile_top_right_45_large_slope = 0x262
-
-tile_bottom_right_gentle_slope_small = 0x289
-tile_bottom_right_gentle_slope_large = 0x28A
-tile_under_bottom_right_gentle_slope_small = 0x2A9
-
-tile_top_right_gentle_slope_small = 0x245
-tile_above_top_right_gentle_slope_small = 0x225
-
-tile_top_right_gentle_slope_large = 0x246
-
-tile_bottom_right_steep_slope_small = 0x22C
-tile_bottom_right_steep_slope_large = 0x24C
-tile_beside_bottom_right_steep_slope_small = 0x22D
-
-tile_top_right_steep_slope_small = 0x287
-tile_beside_top_right_steep_slope_small = 0x288
-
-tile_top_right_steep_slope_large = 0x267
-
-tile_platform_middle = 0x142
-tile_platform_right = 0x143
-
--- Invariant tiles (non-black CRE tiles): leave them unchanged
-if invariant(0, 0) then
-    return true
-end
-
--- Air tiles: blank them out:
-if air(0, 0) then
-    t:set_gfx(tile_number_air, false, false)
-    return true
-end
-
--- Slope tiles: use a matching shape
-if t:type(0, 0) == 1 then
-    bts = t:bts(0, 0) & 0xBF
-    if bts == bts_slope_bottom_right_45 then
-        t:set_gfx(tile_bottom_right_45_slope, bts_hflip(0, 0), false)
-        return true
-    end
-    if bts == bts_slope_bottom_right_45 | 0x80 then
-        t:set_gfx(tile_top_right_45_slope, bts_hflip(0, 0), false)
-        return true
-    end
-    if bts == bts_slope_bottom_right_45_small | 0x80 then
-        t:set_gfx(tile_top_right_45_small_slope, bts_hflip(0, 0), false)
-        return true
-    end
-    if bts == bts_slope_bottom_right_45_large | 0x80 then
-        t:set_gfx(tile_top_right_45_large_slope, bts_hflip(0, 0), false)
-        return true
-    end
-    if bts == bts_slope_bottom_right_steep_small then
-        t:set_gfx(tile_bottom_right_steep_slope_small, bts_hflip(0, 0), false)
-        return true
-    end
-    if bts == bts_slope_bottom_right_steep_small | 0x80 then
-        t:set_gfx(tile_top_right_steep_slope_small, bts_hflip(0, 0), false)
-        return true
-    end
-    if bts == bts_slope_bottom_right_steep_large then
-        t:set_gfx(tile_bottom_right_steep_slope_large, bts_hflip(0, 0), false)
-        return true
-    end
-    if bts == bts_slope_bottom_right_steep_large | 0x80 then
-        t:set_gfx(tile_top_right_steep_slope_large, bts_hflip(0, 0), false)
-        return true
-    end
-    if bts == bts_slope_bottom_right_gentle_small then
-        t:set_gfx(tile_bottom_right_gentle_slope_small, bts_hflip(0, 0), false)
-        return true
-    end
-    if bts == bts_slope_bottom_right_gentle_large then
-        t:set_gfx(tile_bottom_right_gentle_slope_large, bts_hflip(0, 0), false)
-        return true
-    end
-    if bts == bts_slope_bottom_right_gentle_small | 0x80 then
-        t:set_gfx(tile_top_right_gentle_slope_small, bts_hflip(0, 0), false)
-        return true
-    end
-    if bts == bts_slope_bottom_right_gentle_large | 0x80 then
-        t:set_gfx(tile_top_right_gentle_slope_large, bts_hflip(0, 0), false)
-        return true
-    end
-    if bts == bts_slope_whole_bottom_edge and air(0, -1) then
-        t:set_gfx(tile_bottom_edge, false, false)
-        return true
-    end
-    if bts == bts_slope_whole_bottom_edge | 0x80 and air(0, 1) then
-        t:set_gfx(tile_top_edge, false, false)
-        return true
-    end    
-    if (bts == bts_slope_half_bottom_edge_1 or bts == bts_slope_half_bottom_edge_2) and solid(0, 1) then
-        t:set_gfx(tile_half_bottom_edge, false, false)
-        return true
-    end
-    if (bts == bts_slope_half_bottom_edge_1 | 0x80 or bts == bts_slope_half_bottom_edge_2 | 0x80) and solid(0, -1) then
-        t:set_gfx(tile_half_top_edge, false, false)
-        return true
-    end
-    if bts & 0x3F == bts_slope_half_right_edge then
-        t:set_gfx(tile_half_left_edge, not bts_hflip(0, 0), false)
-        return true
-    end
-end
-
--- Solid tiles: look at neighboring edges
-if solid(0, 0) then
-    -- Outside corners (opaque):
-    if outside_right(-1, 0) and inside_left(1, 0) and outside_bottom(0, -1) and inside_top(0, 1) then
-        t:set_gfx(tile_bottom_left_outside_corner, true, false)
-        return true
-    end
-    if outside_left(1, 0) and inside_right(-1, 0) and outside_bottom(0, -1) and inside_top(0, 1) then
-        t:set_gfx(tile_bottom_left_outside_corner, false, false)
-        return true
-    end
-    if outside_right(-1, 0) and inside_left(1, 0) and outside_top(0, 1) and inside_bottom(0, -1) then
-        t:set_gfx(tile_top_left_outside_corner, true, false)
-        return true
-    end
-    if outside_left(1, 0) and inside_right(-1, 0) and outside_top(0, 1) and inside_bottom(0, -1) then
-        t:set_gfx(tile_top_left_outside_corner, false, false)
-        return true
-    end
-
-    -- Horizontal/vertical edges:
-    if outside_right(-1, 0) and inside_left(1, 0) and inside_bottom(0, -1) and inside_top(0, 1) then
-        t:set_gfx(tile_left_edge, true, false)
-        return true
-    end
-    if outside_left(1, 0) and inside_right(-1, 0) and inside_bottom(0, -1) and inside_top(0, 1) then
-        t:set_gfx(tile_left_edge, false, false)
-        return true
-    end
-    if inside_right(-1, 0) and inside_left(1, 0) and outside_bottom(0, -1) and inside_top(0, 1) then
-        t:set_gfx(tile_bottom_edge, false, false)
-        return true
-    end
-    if inside_right(-1, 0) and inside_left(1, 0) and outside_top(0, 1) and inside_bottom(0, -1) then
-        t:set_gfx(tile_top_edge, false, false)
-        return true
-    end
-
-    -- Slope adjacent:
-    if t:type(-1, 0) == 1 and t:bts(-1, 0) & 0x7F == bts_slope_bottom_right_steep_small and inside_left(1, 0) then
-        if not bts_vflip(-1, 0) then
-            t:set_gfx(tile_beside_bottom_right_steep_slope_small, false, false)
-        else
-            t:set_gfx(tile_beside_top_right_steep_slope_small, false, false)
-        end
-        return true
-    end
-    if t:type(1, 0) == 1 and (t:bts(1, 0) & 0x7F) == (bts_slope_bottom_right_steep_small | 0x40) and inside_right(-1, 0) then
-        if not bts_vflip(1, 0) then
-            t:set_gfx(tile_beside_bottom_right_steep_slope_small, true, false)
-        else
-            t:set_gfx(tile_beside_top_right_steep_slope_small, true, false)
-        end
-        return true
-    end
-    
-    if t:type(0, -1) == 1 and t:bts(0, -1) & 0xBF == bts_slope_bottom_right_gentle_small and inside_top(0, 1) then
-        t:set_gfx(tile_under_bottom_right_gentle_slope_small, bts_hflip(0, -1), false)
-        return true
-    end
-    if t:type(0, 1) == 1 and t:bts(0, 1) & 0xBF == bts_slope_bottom_right_gentle_small | 0x80 and inside_bottom(0, -1) then
-        t:set_gfx(tile_above_top_right_gentle_slope_small, bts_hflip(0, 1), false)
-        return true
-    end
-
-
-    if t:type(0, -1) == 1 and (t:bts(0, -1) & 0xBF == bts_slope_bottom_right_45) then
-        t:set_gfx(tile_under_bottom_right_45_slope, bts_hflip(0, -1), false)
-        return true
-    end
-
-    --if t:type(0, 1) == 1 and (t:bts(0, 1) & 0xBF == bts_slope_bottom_right_45 | 0x80) then
-    --    t:set_gfx(tile_under_bottom_right_45_slope, bts_hflip(0, 1), true)
-    --    return true
-    --end
-
-    if inside_right(-1, 0) and inside_left(1, 0) and inside_bottom(0, -1) and inside_top(0, 1) then
-        -- Inside corners:
-        if outside_bottom_right(-1, -1) and not air(1, -1) and not air(-1, 1) and not air(1, 1) then
-            t:set_gfx(tile_bottom_right_inside_corner, false, false)
-            return true
-        end
-        if not air(-1, -1) and outside_bottom_left(1, -1) and not air(-1, 1) and not air(1, 1) then
-            t:set_gfx(tile_bottom_right_inside_corner, true, false)
-            return true
-        end
-        if not air(-1, -1) and not air(1, -1) and outside_top_right(-1, 1) and not air(1, 1) then
-            t:set_gfx(tile_top_right_inside_corner, false, false)
-            return true
-        end
-        if not air(-1, -1) and not air(1, -1) and not air(-1, 1) and outside_top_left(1, 1) then
-            t:set_gfx(tile_top_right_inside_corner, true, false)
-            return true
-        end
-    end    
-
-    -- Platforms:
-    if inside_right(-1, 0) and inside_left(1, 0) and outside_top(0, 1) and outside_bottom(0, -1) then
-        t:set_gfx(tile_platform_middle, false, false)
-        return true
-    end
-    if inside_right(-1, 0) and outside_left(1, 0) and outside_top(0, 1) and outside_bottom(0, -1) then
-        t:set_gfx(tile_platform_right, false, false)
-        return true
-    end
-    if outside_right(-1, 0) and inside_left(1, 0) and outside_top(0, 1) and outside_bottom(0, -1) then
-        t:set_gfx(tile_platform_right, true, false)
-        return true
-    end
-
-    if inside_right(-1, 0) and inside_left(1, 0) and inside_bottom(0, -1) and inside_top(0, 1) then
-        -- Interior
-        t:set_gfx(tile_interior, false, false)
-        return true
-    end    
-
-    -- Other solid tiles: fall back to metal block (to be easy to spot for manual editing)
-    t:set_gfx(tile_unknown_solid, false, false)
-    return true
-end
-
--- Other tiles: mark as unknown (X's)
-t:set_gfx(tile_unknown, false, false)
+../../../Base/Data/RoboTile/RedBrinstar.lua
