+++ Projects/RedBrinstar/Data/RoboTile/RedBrinstar.xml
@@ -1,5 +1,1 @@
-<RoboTileSet>
-<Script Name="Fill" Priority="0">
-require("RedBrinstar")
-</Script>
-</RoboTileSet>
+../../../Base/Data/RoboTile/RedBrinstar.xml
