+++ Projects/WestMaridia/Data/RoboTile/WestMaridiaThick.xml
@@ -1,5 +1,1 @@
-<RoboTileSet>
-<Script Name="Fill" Priority="0">
-require("WestMaridiaThick")
-</Script>
-</RoboTileSet>
+../../../Base/Data/RoboTile/WestMaridiaThick.xml
