+++ Projects/WreckedShip/Data/RoboTile/WreckedShip.xml
@@ -1,5 +1,1 @@
-<RoboTileSet>
-<Script Name="Fill" Priority="0">
-require("WreckedShip")
-</Script>
-</RoboTileSet>
+../../../Base/Data/RoboTile/WreckedShip.xml
